+++ index.html
@@ -6,9 +6,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <title>Кекстаграм</title>
-  <script src="js/function.js"></script>
-  <script src="js/main.js"></script>
+  <title>Кекстаграм</title>   
 </head>
 
 <body>
@@ -235,6 +233,6 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-
+    <script src="./js/main.js" type="module"></script>
 </body>
 </html>
